# server/openapi.yaml
openapi: 3.0.0
info:
  title: EasyCamper API
  version: 1.0.0
  description: API backend EasyCamper (routing, utenti, veicoli, spot, ecc.)
servers:
  - url: http://localhost:5000
    description: Mock server
paths:
  /vehicles:
    get:
      summary: Lista dei veicoli dell'utente
      description: Restituisce la lista dei veicoli dell'utente autenticato. Supporta paginazione e filtri.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Numero massimo di risultati (default 20)
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Offset per la paginazione (default 0)
        - in: query
          name: type
          schema:
            type: string
            enum: [camper, van, autocaravan]
          description: Filtra per tipo veicolo
        - in: query
          name: make
          schema:
            type: string
          description: Filtra per marca
        - in: query
          name: model
          schema:
            type: string
          description: Filtra per modello
        - in: query
          name: year
          schema:
            type: integer
            minimum: 2010
            maximum: 2025
          description: Filtra per anno (solo dal 2010 in poi)
      responses:
        '200':
          description: Array di veicoli
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
        '401':
          description: Autenticazione richiesta
        '403':
          description: Accesso non autorizzato
    post:
      summary: Registra un nuovo veicolo
      description: Registra un nuovo veicolo per l'utente autenticato. Solo tipi camper, van, autocaravan e anno dal 2010 in poi.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVehicle'
      responses:
        '201':
          description: Veicolo creato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Dati non validi (tipo/anno errati)
        '401':
          description: Autenticazione richiesta
        '403':
          description: Accesso non autorizzato
  /vehicles/{vehicleId}:
    get:
      summary: Dettaglio di un veicolo
      description: Restituisce i dettagli di un veicolo posseduto dall'utente autenticato.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: string
          description: ID del veicolo
      responses:
        '200':
          description: Dati del veicolo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Veicolo non trovato o non di proprietà
        '401':
          description: Autenticazione richiesta
        '403':
          description: Accesso non autorizzato
    put:
      summary: Modifica un veicolo esistente
      description: Modifica i dati di un veicolo posseduto dall'utente. Solo tipi camper, van, autocaravan e anno dal 2010 in poi.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: string
          description: ID del veicolo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVehicle'
      responses:
        '200':
          description: Veicolo aggiornato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Dati non validi (tipo/anno errati)
        '404':
          description: Veicolo non trovato o non di proprietà
        '401':
          description: Autenticazione richiesta
        '403':
          description: Accesso non autorizzato
    delete:
      summary: Cancella un veicolo
      description: Cancella il veicolo e tutte le specifiche tecniche correlate (CamperSpec) in modo atomico. Solo se di proprietà dell'utente.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: string
          description: ID del veicolo
      responses:
        '204':
          description: Veicolo e specifiche correlate cancellati
        '404':
          description: Veicolo non trovato o non di proprietà
        '401':
          description: Autenticazione richiesta
        '403':
          description: Accesso non autorizzato
  /spots:
    get:
      summary: Elenco aree di sosta
      parameters:
        - in: query
          name: lat
          required: true
          schema:
            type: number
          description: Latitudine utente
        - in: query
          name: lng
          required: true
          schema:
            type: number
          description: Longitudine utente
        - in: query
          name: filters
          required: false
          schema:
            type: object
          description: Filtro servizi/disponibilità
      responses:
        '200':
          description: Array di spot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Spot'
  /spots/{spotId}:
    get:
      summary: Dettaglio di un’area di sosta
      parameters:
        - in: path
          name: spotId
          required: true
          schema:
            type: string
          description: ID dello spot
      responses:
        '200':
          description: Dati dello spot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spot'
        '404':
          description: Spot non trovato
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Vehicle:
      type: object
      properties:
        id:
          type: string
          example: "abc123"
        userId:
          type: string
          example: "user456"
        type:
          type: string
          enum: [camper, van, autocaravan]
        make:
          type: string
          example: "Fiat"
        model:
          type: string
          example: "Ducato"
        height:
          type: number
          description: Altezza in metri
          example: 3.0
        length:
          type: number
          description: Lunghezza in metri
          example: 7.2
        weight:
          type: number
          description: Peso in tonnellate
          example: 3.5
        imageUrl:
          type: string
          format: uri
        year:
          type: integer
          minimum: 2010
          maximum: 2025
          example: 2020
      required:
        - id
        - userId
        - type
        - make
        - model
        - year
    NewVehicle:
      type: object
      properties:
        type:
          type: string
          enum: [camper, van, autocaravan]
        make:
          type: string
        model:
          type: string
        height:
          type: number
        length:
          type: number
        weight:
          type: number
        year:
          type: integer
          minimum: 2010
          maximum: 2025
      required:
        - type
        - make
        - model
        - year
    Spot:
      type: object
      properties:
        id:
          type: string
          example: "spot789"
        name:
          type: string
          example: "Area di sosta Lago Maggiore"
        coords:
          type: object
          properties:
            lat:
              type: number
              example: 45.9
            lng:
              type: number
              example: 8.5
        services:
          type: array
          items:
            type: string
            example: "acqua"
        availability:
          type: string
          enum: [free, occupied]
        rating:
          type: number
          format: float
          example: 4.2
        images:
          type: array
          items:
            type: string
            format: uri
      required:
        - id
        - name
        - coords
...